requirements.txt
flask
flask-script
flask-WTF

pillow

gunicorn
